\name{emk}
\alias{emk}
\title{Empirical cumulants for individual data}
\usage{
emk(DATA, order = 1:4)
}
\arguments{
  \item{DATA}{a vector of individual data}

  \item{order}{order of the cumulant. Must be between 1 and
  4.}
}
\value{
A vector of cumulants
}
\description{
Empirical cumulants for individual data
}
\author{
François Pelletier
}

