\name{empCF}
\alias{empCF}
\title{Empirical Characteristic function}
\usage{
empCF(t, DATA)
}
\arguments{
  \item{t}{Vector of transform variates}

  \item{DATA}{Vector of individual data}
}
\value{
A vector of CF values
}
\description{
Evaluate the empirical CF from a data sample
}
\author{
François Pelletier
}

