\name{empCGF}
\alias{empCGF}
\title{Empirical Coment Generating function}
\usage{
empCGF(t, DATA)
}
\arguments{
  \item{t}{Vector of transform variates}

  \item{DATA}{Vector of individual data}
}
\value{
A vector of CGF values
}
\description{
Evaluate the empirical CGF from a data sample
}
\author{
François Pelletier
}

