\name{empMGF}
\alias{empMGF}
\title{Empirical Moment Generating function}
\usage{
empMGF(t, DATA)
}
\arguments{
  \item{t}{Vector of transform variates}

  \item{DATA}{Vector of individual data}
}
\value{
A vector of MGF values
}
\description{
Evaluate the empirical MGF from a data sample
}
\author{
François Pelletier
}

